refactor(*): rename "extension" -> "enhancer"
