+++ src/enhancers.js
@@ -1,0 +1,2 @@
+export * from './enhancers/withNoNewKeyword'
+export * from './enhancers/withGlobalDefaults'
