+++ src/enhancers/withNoNewKeyword.js
