--- src/extensions.js
@@ -1,2 +1,0 @@
-export * from './extensions/withNoNewKeyword'
-export * from './extensions/withGlobalDefaults'
